Add base_url and minor fixes

+++ savida/fixture.py
@@ -10,15 +10,20 @@ class ServerWrapper(object):
         self._proc = None
         self.server = server
 
+    @property
+    def base_url(self):
+        return self.server.base_url
+
     def start(self):
         self._proc = multiprocessing.Process(target=self.server.start)
         self._proc.start()
         return self
 
     def stop(self):
-        self._proc.terminate()
-        self._proc.join()
-
+        if self._proc is not None:
+            self._proc.terminate()
+            self._proc.join()
+        self._proc = None
 
 @contextlib.contextmanager
 def http_server(*args, **kwargs):
